+++ components/create/AddQuestionModal/AddQuestionForm.jsx
@@ -1,0 +1,50 @@
+import React, { useState, useEffect } from "react";
+import styles from '../../../styles/Create.module.css';
+const {backdrop, modal} = styles;
+
+export default function AddQuestionForm({closeModal}) {
+  const submitQuestion = async (e) => {
+    console.log({e})
+    e.preventDefault();
+  //   const data = {
+  //    triviaId: 2,
+  //    roundNum: 1,
+  //    questionNum: 2,
+  //    content: 'Who is the best partner in the world?',
+  //    type: 'text',
+  //    correctAnswer: 'Daniel'
+  // }
+  // await fetch('http://localhost:3000/api/post/questions', {
+  //    method: 'PUT',
+  //    headers: {
+  //       'Content-Type': 'application/json'
+  //    },
+  //    body: JSON.stringify(data)
+  //  })
+ }
+  return (
+    <div className={backdrop}>
+    <div className={modal}>
+     <form>
+         <div>
+           <label for="question">Question</label>
+           <input type="text" name="question"></input>
+         </div>
+         <div>
+           <label for="answer">Answer</label>
+           <input type="text" name="answer"></input>
+         </div>
+
+      <div>
+        <button type="button" onClick={closeModal}>
+           Close the goddamm Modal
+        </button>
+      </div>
+      <div>
+        <button onClick={submitQuestion}>Submit Question</button>
+      </div>
+     </form>
+    </div>
+  </div>
+  )
+}
