+++ components/create/AddQuestionModal/AddQuestionModal.jsx
@@ -1,0 +1,14 @@
+import { useRef, useEffect, useState } from 'react';
+import { createPortal } from 'react-dom';
+
+export default function AddQuestionModal({ children, selector }) {
+  const ref = useRef();
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => {
+    ref.current = document.querySelector(selector);
+    setMounted(true);
+  }, [selector]);
+
+  return mounted ? createPortal(children, ref.current) : null;
+}
